+++ packages/docs/site/src/routes/compounds/time_region_value/index.svelte
@@ -2,7 +2,7 @@
 	import Index from '@svizzle/time_region_value/src/routes/index.svelte';
 
 	import {_groups, _yearRange} from './_data/stores';
-	import {hrefBase} from './_navigation';
+	import {hrefBase} from './_config';
 </script>
 
 <Index